ai-cooking-agent/requirements.txt
agent-framework-core==1.0.0rc3
openai>=1.99.9
python-dotenv>=1.0.1
debugpy>=1.8.14
